moved Checkbox back to common package
