+++ jdi-light-bootstrap-tests/src/main/java/io/github/com/sections/form/CheckboxesAndRadiosWithoutLabels.java
@@ -2,7 +2,7 @@ package io.github.com.sections.form;
 
 import com.epam.jdi.light.elements.composite.Section;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
-import com.epam.jdi.light.ui.bootstrap.elements.complex.Checkbox;
+import com.epam.jdi.light.ui.bootstrap.elements.common.Checkbox;
 import com.epam.jdi.light.ui.bootstrap.elements.common.RadioButton;
 
 public class CheckboxesAndRadiosWithoutLabels extends Section {
