+++ jdi-light-bootstrap-tests/src/main/java/io/github/com/sections/form/CheckboxesDefaultInline.java
@@ -1,7 +1,7 @@
 package io.github.com.sections.form;
 
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
-import com.epam.jdi.light.ui.bootstrap.elements.complex.Checkbox;
+import com.epam.jdi.light.ui.bootstrap.elements.common.Checkbox;
 
 public class CheckboxesDefaultInline {
