+++ jdi-light-bootstrap-tests/src/main/java/io/github/com/sections/form/FormCustomStyles.java
@@ -3,7 +3,7 @@ package io.github.com.sections.form;
 import com.epam.jdi.light.elements.composite.Form;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
 import com.epam.jdi.light.ui.bootstrap.elements.common.Button;
-import com.epam.jdi.light.ui.bootstrap.elements.complex.Checkbox;
+import com.epam.jdi.light.ui.bootstrap.elements.common.Checkbox;
 import com.epam.jdi.light.ui.bootstrap.elements.common.TextField;
 import io.github.com.entities.FormContacts;
