+++ jdi-light-bootstrap-tests/src/main/java/io/github/com/sections/form/FormDisabled.java
@@ -4,7 +4,7 @@ import com.epam.jdi.light.elements.complex.Selector;
 import com.epam.jdi.light.elements.composite.Section;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
 import com.epam.jdi.light.ui.bootstrap.elements.common.Button;
-import com.epam.jdi.light.ui.bootstrap.elements.complex.Checkbox;
+import com.epam.jdi.light.ui.bootstrap.elements.common.Checkbox;
 import com.epam.jdi.light.ui.bootstrap.elements.common.TextField;
 import com.epam.jdi.light.ui.bootstrap.elements.complex.MediaObject;
