+++ jdi-light-bootstrap-tests/src/main/java/io/github/com/sections/form/FormInlinePreference.java
@@ -4,7 +4,7 @@ import com.epam.jdi.light.elements.complex.dropdown.DropdownSelect;
 import com.epam.jdi.light.elements.composite.Form;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
 import com.epam.jdi.light.ui.bootstrap.elements.common.Button;
-import com.epam.jdi.light.ui.bootstrap.elements.complex.Checkbox;
+import com.epam.jdi.light.ui.bootstrap.elements.common.Checkbox;
 import io.github.com.entities.FormContacts;
 
 public class FormInlinePreference extends Form<FormContacts> {
