+++ jdi-light-bootstrap/src/main/java/com/epam/jdi/light/ui/bootstrap/asserts/CheckboxAssert.java
@@ -3,7 +3,7 @@ package com.epam.jdi.light.ui.bootstrap.asserts;
 import com.epam.jdi.light.asserts.generic.SelectedAssert;
 import com.epam.jdi.light.asserts.generic.UIAssert;
 import com.epam.jdi.light.common.JDIAction;
-import com.epam.jdi.light.ui.bootstrap.elements.complex.Checkbox;
+import com.epam.jdi.light.ui.bootstrap.elements.common.Checkbox;
 
 
 import static com.epam.jdi.light.asserts.core.SoftAssert.jdiAssert;
